adapt fe to new municipal kpi endpoint
